pyproject.toml
[tool.poetry]
name = "pq-engine"
version = "0.1.0"
description = "pseudo quantum engine"
authors = ["bbrfkr <bbrfkr@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pq_engine"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"
cupy-cuda12x = "^12.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79

[tool.mypy]
ignore_missing_imports = true
